requirements.txt
# Core dependencies
openai>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.1.0
transformers>=4.40.0
torch>=2.3.0           # Latest version for better Metal support on Apple Silicon
huggingface-hub>=0.20.0
ctransformers>=0.2.28  # For faster inference with GGUF models (with Apple Metal support)
hf_transfer>=0.1.0     # For faster model downloads
accelerate>=0.21.0     # For better GPU offloading

# Web search and scraping
duckduckgo-search>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0

# FastAPI for web interface
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Additional utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
rich>=13.0.0
streamlit>=1.28.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
